Keep value diagnostics in value package
